successfully scraped url into homepage, need to sort and implement data
