--- src/android/GetGrinnected/MyApplication/app/src/main/java/com/example/myapplication/api/Condition.kt
@@ -1,7 +1,0 @@
-package com.example.myapplication.api
-
-data class Condition(
-    val code: Int,
-    val icon: String,
-    val text: String
-)
