--- src/android/GetGrinnected/MyApplication/app/src/main/java/com/example/myapplication/api/Constant.kt
@@ -1,5 +1,0 @@
-package com.example.myapplication.api
-
-object Constant {
-    val apiKey = "235ac1fa18c943eba0632820250804"
-}
