--- src/android/GetGrinnected/MyApplication/app/src/main/java/com/example/myapplication/api/DAta.kt
@@ -1,3 +1,0 @@
-package com.example.myapplication.api
-
-class DAta : ArrayList<DAtaItem>()
