--- src/android/GetGrinnected/MyApplication/app/src/main/java/com/example/myapplication/api/Location.kt
@@ -1,12 +1,0 @@
-package com.example.myapplication.api
-
-data class Location(
-    val country: String,
-    val lat: Double,
-    val localtime: String,
-    val localtime_epoch: Int,
-    val lon: Double,
-    val name: String,
-    val region: String,
-    val tz_id: String
-)
