--- src/android/GetGrinnected/MyApplication/app/src/main/java/com/example/myapplication/api/Weather.kt
@@ -1,6 +1,0 @@
-package com.example.myapplication.api
-
-data class Weather(
-    val current: Current,
-    val location: Location
-)
