--- src/android/GetGrinnected/MyApplication/app/src/main/java/screens/test.kt
@@ -1,4 +1,0 @@
-package screens
-
-class test {
-}
