+++ src/android/hello.kts
@@ -1,4 +1,7 @@
 import java.net.URL
+import java.io.BufferedReader
+import java.io.InputStreamReader
+import javax.net.ssl.HttpsURLConnection
 
 // Test1(first) -> String
 // first - String
@@ -63,20 +66,49 @@ fun add(a: Int, b: Int): Int {
     return sum
 }
 
+fun getDataFromUrl(url: String): String? {
+    var connection: HttpsURLConnection? = null
+    return try {
+        val urlObj = URL(url)
+        connection = urlObj.openConnection() as HttpsURLConnection
+        connection.requestMethod = "GET"
 
-fun readStringFromURL(urlString: String): String {
-    val url = URL(urlString)
-    return url.readText()
+        val responseCode = connection.responseCode
+        if (responseCode == HttpsURLConnection.HTTP_OK) {
+            val reader = BufferedReader(InputStreamReader(connection.inputStream))
+            val response = StringBuilder()
+            var line: String?
+            while (reader.readLine().also { line = it } != null) {
+                response.append(line)
+            }
+            reader.close()
+            response.toString()
+        } else {
+            null
+        }
+    } catch (e: Exception) {
+        e.printStackTrace()
+        null
+    } finally {
+        connection?.disconnect()
+    }
 }
 
+
+
+
 fun main(){
-    val content = readStringFromURL("http://node16049-csc324--spring2025.us.reclaim.cloud:11014/")
-    println(content)
+    val url = "https://node16049-csc324--spring2025.us.reclaim.cloud/events"
+    val result = getDataFromUrl(url)
+    if (result != null) {
+        println(result)
+    } else {
+        println("Failed to get data from URL")
+    }
     println("Hello World")
     test1("Ethan")
     println(test2(7))
     println(recursive(1))
-    val result = add(5, 3)
     println(result)
     loop()
     val trialRun = arrayOf(" learn Kotlin ", " sleep ", " suffer ")
